pyproject.toml
[project]
name = "trackers"
readme = "README.md"
dependencies = [
    "aioauth-client",
    "aiohttp",
    "aiohttp_session[secure]",
    "aiomsgpack",
    "aionotify",
    "aniso8601",
    "arsenic",
    "asynctest",
    "beautifulsoup4",
    "calmjs.parse",
    "dulwich",
    "fixtures",
    "geographiclib",
    "htmlwrite",
    "jsonpointer",
    "libsass",
    "more-itertools",
    "msgpack",
    "numpy",
    "nvector==0.7.6",
    "polyline",
    "python-datauri",
    "python-dateutil",
    "python-slugify",
    "pyyaml",
    "tap.py",
    "testresources",
    "testscenarios",
    "uvloop",
]

[project.optional-dependencies]
dev = [
    "flake8",
    "black",
    "isort",
]

[project.scripts]
add_gpx_to_event_routes = "trackers.bin_utils:add_gpx_to_event_routes"
assign_rider_colors = "trackers.bin_utils:assign_rider_colors"
convert_to_static = "trackers.bin_utils:convert_to_static"
gen_key = "trackers.auth:gen_key"
load_riders_from_csv = "trackers.bin_utils:load_riders_from_csv"
print_names_and_trackers = "trackers.bin_utils:print_names_and_trackers"
process_event_routes = "trackers.bin_utils:process_event_routes"
reformat_event = "trackers.bin_utils:reformat_event"
run_analyse = "trackers.bin_utils:analyse"
run_qunit_tests = "trackers.client_test_tools:qunit_runner"
serve = "trackers.serve:main"
store_analyse = "trackers.bin_utils:store_analyse"
trackleaders_get_config = "trackers.sources.trackleaders:get_config"
update_bounds = "trackers.bin_utils:update_bounds"

[tool.pdm]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
